// src/components/types/baseTypes.ts
export interface TeamStats {
    totalGold: number;
    inhibitors: number;
    towers: number;
    barons: number;
    totalKills: number;
    dragons: string[];
    participants: WindowParticipant[];
}

export interface Tournament {
    id: string;
}

export interface Window {
    esportsGameId: string;
    esportsMatchId: string;
    gameMetaData: GameMetadata;
    frames: WindowFrame[]
}

export interface WindowFrame {
    blueTeam: TeamStats;
    gameState: "in_game" | "finished";
    rfc460Timestamp: string;
    redTeam: TeamStats;
}

export interface WindowParticipant {
    participantId: number;
    totalGold: number;
    level: number;
    kills: number;
    deaths: number;
    assists: number;
    creepScore: number;
    currentHealth: number;
    maxHealth: number;
}